+++ src/services/databaseservice.h
@@ -14,6 +14,12 @@ public:
     static bool setupTables();
     static QString getAppData(QString name);
     static bool setAppData(QString name, QString value);
+    static bool reinitializeDiskDatabase();
+private:
+    static QString getDiskDatabasePath();
+    static bool removeDiskDatabase();
+    static bool createMemoryConnection();
+    static bool createDiskConnection();
 };
 
 #endif // DATABASESERVICE_H