+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.67
 - fixed a problem with storing todo items with a reminder and an empty description
+- added a button to reinitialize the database to the settings dialog
 
 ## 0.66
 - fixed a problem with storing wrong todo list urls in the settings